+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ...
 
+## v0.6.0
+
+- Workflows can define a staticmethod `applies_to` that takes an instance and returns whether or not the workflow applies to it.
+
+## v0.5.1
+
+- Prevent re-submitting Done tasks.
+
+## Previous versions
+
+...
+
 ## v0.3.1
 
 - Changed default ordering for tasks: more recent first